Implement product inventory module with WAC
Add complete inventory system with SQLite-backed ledger + snapshot architecture. Implements Weighted Average Cost (WAC) calculations, offline-first sync strategy, and per-product transaction history. Replaces in-memory product data with append-only transaction log. Changes product IDs from int to String (UUIDs) for global identity. Includes comprehensive technical documentation in docs/product_inventory_module.md. Integrates with ProductBloc, sales/purchases, and all related presentation layers.

+++ lib/presentation/products/widgets/product_info_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:trucky/core/constants/font_constants.dart';
 import 'package:trucky/core/utils/number_formater.dart';
 import 'package:trucky/core/utils/widget_extensions.dart';
-import 'package:trucky/presentation/products/bloc/product_bloc.dart';
+import 'package:trucky/presentation/products/bloc/product_models.dart';
 import 'package:trucky/presentation/products/widgets/product_list.dart';
 import 'package:trucky/presentation/widgets/label_widget.dart';
 
@@ -19,7 +19,7 @@ class ProductInfoWidget extends StatelessWidget {
     this.currency = '',
   });
 
-  int get availableStock => selectedProduct.availableStock;
+  double get availableStock => selectedProduct.availableStock;
 
   @override
   Widget build(BuildContext context) {
